+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.0.4 - 2019-04-26
 
 * Handle cases where compression is done before signing.
-* Add support for additional encryption algorithms
+* Add support for additional encryption algorithms.
+* Use binary encoding for encryption and signatures.
 * Remove support for Python 2.
 
 ## 1.0.3 - 2018-05-01